Cherrypick a bugfix for clang.
This fix is necessary to support clang-style FORTIFY on ChromeOS.
Specifically, it fixes an issue with evaluating enable_if attributes on
variadic functions, such as snprintf.

As with all patches, this isn't expected to cause a regression. If it
does break something, said thing will be related to clang's enable_if
attribute, which is *very* rarely used outside of FORTIFY.

BUG=None
TEST=building x86-generic with clang (and the FORTIFY patch) works

Change-Id: Ie32792bf49d3f66ffcd1d8a1142bad86fc66019c
Reviewed-on: https://chromium-review.googlesource.com/368620
Commit-Ready: George Burgess <gbiv@google.com>
Tested-by: George Burgess <gbiv@google.com>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>