media-libs/mesa-img: Consolidate patches
Consolidate patches in preparation for an update release of the DDK.
Most of the 010x patches have landed in the upstream DDK, so are now
integrated into the main "Introduce PowerVR driver" patch.

Also take this opportunity to clean up the patches, and fix some
build warnings.

BUG=chrome-os-partner:56148
TEST=Boot to UI.  Run WebGL aquarium.  Run a CTS test.

Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Change-Id: I646d099bc37f87fabf74964cdab4bebeb4c313f7
Reviewed-on: https://chromium-review.googlesource.com/368660
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>