media-libs/arc-mesa-img: publish pretend 24 bit depth support
Publish a config that advertises support for 24-bit depth buffers.
When the app slects this config, the DDK will basically ignore this
value and will create a (32-bit) depth buffer.

This allows dEQP to run without skipping tests due to missing configs.
But dEQP on Android should really be fixed so that it can run on systems
that don't provide 24-bit depth buffer configs.  (This is already the
case on surfaceless.)

BUG=chrome-os-partner:55634
TEST=ARC++ dEQP.  Examine logcat to make sure tests are actually running.

Change-Id: I7186c1b3a7926cf33e1dd8c687c62e07ec828c96
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/371198
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>