ebuild: add cr50 updater startup files
The usb_updater utility needs to make sure that the CR50 device is
running the latest firmware included in the Chrome OS image. Once
firmware update happens the Chromebook needs to be restarted for the
latest code to take effect.

Various exit codes generated by usb_updater indicate various update
results. Since the upstart framework does not allow for non-zero exit
values within a single script, two scripts are required, one to run
the updater, and another one to act on the update result.

Exit code of 1 or 2 should cause a system restart.

The update process state is also saved in a file to ensure that there
is no rolling reboots if something goes wrong, maximum number of
reboots for a single cr50 firmware image is 4.

The trunksd daemon needs to be held off until the update result is
available, this makes sure that there is no competition in accessing
/dev/tpm0. A customized version of the upstart config is required for
the cr50 case.

CQ-DEPEND=CL:376301
BUG=chrome-os-partner:49954
TEST=loaded old cr50 version on a kevin-tpm2 device and installed the
     new chrome os image on it. Observed the setup rebooting twice,
     verified in the end that the cr50 had the latest RO and RW and
     that trunksd is running.

Change-Id: Id3d2db3092f3076c56c7a852d0259658489899f2
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/375960
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>