net-print/cups: Patch to be strict on filter names in PPDs
We require that all filters specified in PPDs are just the program name,
and do not contain any relative or absolute path info.

BUG=chromium:633384
TEST=Manually ran cuptestppd on device with dummy PPD. Made sure it
could pass with a valid filter, but when usingn path info (adding . or /
to filter), it would fail. Tested with both cupsFilter and cupsFilter2.

Change-Id: I891e6ed9cba6d262d6db68b0603069bc52779919
Reviewed-on: https://chromium-review.googlesource.com/376021
Commit-Ready: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>