llvm: use epatch to get upstream patch.
We use git cherry-pick to get upstream patch. This would change the
git hash of the git repository for each build. This causes gomma update
the LLVM even we did nothing to it.
We then uses git cherry-pick -n to workaround it. But it is better to use
epatch for it.

BUG=chromium:626852
TEST=sudo ebuild llvm.

Change-Id: I3fd54a9b4a1198cfa4a9d278228096f2ffdf4250
Reviewed-on: https://chromium-review.googlesource.com/376205
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>