config.kunimitsu: Add FSP 2.0 support
kunimitsu can be used with either FSP1.1 or FSP2.0, hence
add a kunimitsu config with required configs for supporting
FSP 2.0.
In order to select between the 2 FSPs add a USE switch
in coreboot ebuild which is used for selecting appropriate config.
Also update eclass for building chromeos-ec and pd for kunimitsu.

Change-Id: I5fc575722925369a7e791b03dbe36a3a51140e39
Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/379977
Reviewed-by: Aaron Durbin <adurbin@chromium.org>