syslinux: drop local copy
We have a newer version in portage-stable now we want to use.

Keep in mind that syslinux is not used on any official Chrome OS device
as we have our own custom firmware everywhere.  It is only used when
booting Chromium OS.

Our copy had a few custom changes:
- Turn off the video cursor before booting the kernel
- Sometimes disable display of the syslinux banner when booting
- Disable display of all info messages

The first patch is no longer needed: our boot splash code takes care
of turning off the cursor before showing the screen, either directly
(when USE=vtconsole), or by letting frecon do it when it creates the
new VT.

Disabling the boot banner happens only for some boot paths (e.g. the
isolinux code), but not all (e.g. the serial port), and it disables
only part of the banner.  It also doesn't disable all messages that
come up early on.  In the end, it doesn't make things all the much
more quiet to the point where we need to maintain it.

The code also tried to disable the common writemsg function, but it
is only called for info/debug messages, none of which are displayed
by default.  Which means it's kind of a pointless change.

By dropping our custom syslinux version, we make the output a little
bit more verbose for Chromium OS builds, but since we aren't silent
now, it's not a big deal.

BUG=chromium:508713
TEST=`emerge syslinux` uses the new one in portage-stable
TEST=vmtests pass
CQ-DEPEND=CL:382771

Change-Id: I922710079ac5f4c9ca9c177eab880a99ada1415e
Reviewed-on: https://chromium-review.googlesource.com/382772
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>