chromeos-chrome: reduce debug file size for release build on 32-bit
The size of .dwp file exceeds 4GB, so we add -femit-struct-debug-reduced
to reduce the size of debug file. With this flag, the size of debug
file is reduced and the symbol files remains the same.

We are still considering other options (current solution may cause some
issues for debugger users. Although it is not clear how many people
actually do that with the ebuild workflow).

BUG=chromium:641188
TEST=the size of .dwp file is reduced by half. The generated symbol
     files remains the same.

Change-Id: I49ecd0bfb12c8f41a08ba0a4420d26cd3a861604
Reviewed-on: https://chromium-review.googlesource.com/381781
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>