pnm2png: split into dedicated package
This is the only thing that we've done custom to the libpng ebuild, and
it's keeping us from easily upgrading to upstream unmodified.  Since the
code doesn't really need to be integrated into the libpng ebuild, and it
rarely changes (1.2.56 to 1.6.25 is entirely style updates), just create
a dedicated ebuild to hold it.  This will let us drop our custom version
of libpng from our repo and move to latest upstream.

BUG=chromium:599917
TEST=precq passes (clean upgrade of libpng/pnm2png/lorgnette)

Change-Id: I2b31dcee4907509392cdab3ceac0909d51857063
Reviewed-on: https://chromium-review.googlesource.com/385596
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jungshik Shin <jshin@chromium.org>