Update FreeType to 2.7 with local changes
Disable "subpixel hinting" for now. This patch will be dropped to keep
the default value (2) for TT_CONFIG_OPTION_SUBPIXEL_HINTING
after merging to a branch.

Update 'enable-valid' patch to 2.7. This may be dropped in the next
iteration because Harfbuzz validates OT downstream (and more
thoroughly). Keep it now for merging to a branch (to minimize changes).

Besides, apply the difference between 2.7 and c38be52bf8de3

BUG=chromium:646635
TEST=emerge-{x86-alex,amd64-generic,daisy} freetype succeeds.
TEST=cbuildbot chromiumos-sdk
TEST=cbuildbot  amd64-generic-full x86-generic-full arm-generic-full
TEST=manual/visual: WebUI rendering and web page rendering do not have any noticeable regression.

Change-Id: Iec2873422d6309b12294f4bbf60b7e95184c7432
Reviewed-on: https://chromium-review.googlesource.com/385241
Commit-Ready: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>