dev-lang/go: upgrade to go-1.7.1.
This ebuild brings Go 1.7 to Chrome OS (both host and cross compilers).
For details on testing and known issues, see http://go/Go1.7-On-ChromeOS

Most of the changes are in the implementation of the toolchain, runtime,
and libraries. There is one minor change to the language specification.
For more details, see https://golang.org/doc/go1.7

BUG=None
TEST='emerge cross-x86_64-cros-linux-gnu/go' and ran compiler and standard library tests on a panther device.
TEST='emerge cross-i686-pc-linux-gnu/go' and ran compiler and standard library tests on a x86-zgb device.
TEST='emerge cross-armv7a-cros-linux-gnueabi/go' and ran compiler and standard library tests on a daisy device.

Change-Id: I6230c0dde31fb2f4b6e57edfb9375870fc6b01dc
Reviewed-on: https://chromium-review.googlesource.com/386282
Commit-Ready: Rahul Chaudhry <rahulchaudhry@chromium.org>
Tested-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>