make.conf.sdk-chromeos: new fragment to hold internal SDK settings
The chromeos-overlay/make.conf file currently impacts both boards and
the sdk.  This hasn't been a problem previously as we had very little
in there, but as we try to add more customizations for boards, we're
making it problematic for the sdk.  Start a new file just for the sdk
so we avoid such problems.

BUG=chromium:440498
TEST=updating an sdk has same settings, as does creating a new sdk

Change-Id: Ia63c63fad496d8a97bb613c44a760950e513f271
Reviewed-on: https://chromium-review.googlesource.com/391592
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>