chromeos-kernel-4_4: correct ebuild name
This is not a correct ebuild version/name. It was generated because a
bad tag was manually pushed into the kernel.git repo (but subsequently
was deleted). This should the CQ uprev process back in working order.

We need to delete the bad tag in 'chromeos-version.sh' because bots may
have already propagated the tag, and tags don't get transitively
deleted.

BUG=chromium:653672
TEST=emerge-${BOARD} chromeos-kernel-4_4

Change-Id: Iea5dd194ff84aab3113d1229a299973c86e8a5b7
Fixes: 07213432b36d ("Marking set of ebuilds as stable")
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/394752
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>