arc-build: add pkgconfig-arc to help build arc-mesa
Upstream wants the Android libs that are required when building
arc-mesa to be found using pkg-config.  To do this, let's create
the .pc files we need and put them in the Android sysroot.
In addition, we need another version of pkg-config for amd64
Android builds.

BUG=b:32077475
TEST=emerge-cyan arc-mesa succeeds
CQ-DEPEND=CL:403315

Change-Id: I7d0538dc09109d1deaeb584e61ee9febf4b5f902
Reviewed-on: https://chromium-review.googlesource.com/403237
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>