cros-kernel2: add basic rockchip64 config support
Upstream has most basic support for rk3399, so with a little creative
work (e.g., stripping down our .dtsi files a bit, and disabling SELinux
enforcement), we can boot to console with it.

The defconfig was generated by applying our local config to the
upstream and accepting all the defaults. Works well enough for a start!

BUG=none
TEST=build upstream for kevin, boot to console

Change-Id: Ic1fcc82bb6b5d55b68121b3d88ab36c93bf0b90b
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/411365
Reviewed-by: Douglas Anderson <dianders@chromium.org>