Install touch seccomp policies per-architecture
Previously there was only 1 set of seccomp policies for the touch
firmware updater scripts, but now we have some of these devices on
multiple architectures.  As a result, they use different syscalls
if they're on arm, vs x86, etc so the policies to carry over.
This CL changes the ebuild to pull policies from a directory named
with the architecture in question.

BUG=chrome-os-partner:60015
TEST=emerged and deployed touch_updater on several systems that use
these seccomp policies on different architectures.  The policies are
pulled from the right directories and work smoothly
CQ-DEPEND=CL:412322

Change-Id: I52d269a57e7b30159af25a1a1a979c285d319b46
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/412340
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>