cr50: separate script and application error codes
The usb_updater application which is used to update CR50 firmware
could return exit code of 3 in case something goes wrong. It is worth
trying the update again in that case (the maximum number of attempts
is still limited by the cr50-update.conf script).

Problems reported by the cr50-update.conf script are such that there
is no point in retrying (unknown board, or non-existing binary). They
should be reported as a different return code and not lead to the
device reboot.

BRANCH=reef-factory
BUG=none
TEST=verified that cr50 firmware update is still functional on a reef
     device.

Change-Id: I8bf9f79e60baba1da725fbb1cca18b0f28ee2e40
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/413587
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>