Make authpolicy a dependency of Chrome
The authpolicy package handles authentication and policy fetch from Active
Directory servers. Since Chrome is the only package that talks to it, it is
added as a dependency of Chrome.

Authpolicy is disabled by default. During an initial testing phase it is planned
to be enabled on certain boards only.

BUG=chromium:667696
TEST=Emerges, can be turned on via board-specific overlays

Change-Id: I32e1bb61e28fe4e35c29a64d48d0d479e472d0c9
Reviewed-on: https://chromium-review.googlesource.com/413985
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>