arc-mesa: Add LLVM runtime for ARC++ to SDK
The prebuilts from arc-llvm do not provide us with the necessary
headers and tools to use llvmpipe for arc-mesa (libLLVM, llvm-config
and libLLVM headers).
Add scripts and ebuilds to pick up the necessary files from an ARC++
build and install them in the SDK.

The only user is Mesa, so we create a new package arc-llvm-mesa. We
will revisit that for N since the way the ARC++ toolchain is managed has
changed.

Instructions to build the LLVM runtime files are available in the
gather.sh script.

Also tweak the arc-mesa ebuild to use that version of the runtime.

Still TODO: using pkg-config should allow us to clean up the Mesa
configure script.
Still TODO: revisit for N-based version of ARC++.

BUG=b:28802929
TEST=Run gather.sh script and upload tar.gz file.
TEST=./build_packages automatically emerges arc-llvm-mesa and adds
libLLVM.so and friends to /opt/android/arc-llvm-mesa/.
TEST=amd64-generic-cheets image boots in qemu and uses llvmpipe.

Change-Id: I2afd17dfbfdcbf3987d5c8f0ae9015f85a515479
Reviewed-on: https://chromium-review.googlesource.com/414586
Commit-Ready: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>