ec-utils: remove build products before running make
When ec-utils 9999 ebuild is emerged, the entire contents of the local
tree is copied into the sandbox.

If the local tree contains built images, make invoked by ebuild
reports that the target is up to date. This is very harmful when board
the ebuild is invoked for is not an x86 board: the generated image is
include in the root file system, but it is not executable on the
target.

Invoking 'make clean' first makes sure that the image is rebuilt.

BUG=none
TEST=proper executable for gru is generated even when prebuilt
     usb_updater for x86 exists.

Change-Id: Id75b5efc658b3dbae26dd3037c795c6bc28a91a7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/414308
Reviewed-by: Shawn N <shawnn@chromium.org>