Update evtest to version 1.31
This patch updated evtest to version 1.31.  This version includes
some extra code that prints out the values for INPUT_PROP_DIRECT and
INPUT_PROP_POINTER when it's displaying the evdev device's properties
at start.  These properties were invisible before, but are instrumental
in the way that Chrome decides which touch devices are touchpads and
which are touchscreens.  Therefore, adding them to this test tool will
make it easier to debug any issues that arise from mis-set values in
these two parameters in touch devices.

Note, there are a few changes to the ebuild that seem random.  The
new ebuild changes were copied from Gentoo here:
https://gitweb.gentoo.org/repo/gentoo.git/tree/app-misc/evtest/evtest-1.31.ebuild
It is the ebuild for 1.31 suggested on there and fixes some naming
issues with the newest mirrors tarball files.

BUG=chromium:671345
TEST=emerged and tested on a Chromebook -- the new version built
correctly and was functional on the device I tested it on.  Not only
showing the existing functionality but also reporting the new properties
as well.

Change-Id: Id1746376fe54e97c22356b1691fcbace1719fc27
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/416500
Reviewed-by: Mike Frysinger <vapier@chromium.org>