media-libs/[arc-]mesa-img: update patch adding PVR DRI driver
The DRI driver server_wait_sync fence extension function can be called with
a NULL fence when a reusable sync is given to eglWaitSyncKHR. This would
cause a segfault in the PVR DRI driver as it wasn't handling this case. Fix
this by adding a fence NULL check.

BUG=chrome-os-partner:60359
TEST=Run dEQP-EGL.functional.reusable_sync.valid.wait_server

Change-Id: I4f9aac848071e432d2d6bf29810d90e8de7f92ce
Signed-off-by: Frank Binns <frank.binns@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/417322
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>