Add authpolicy dependency to target-chromium-os ebuild
This CL is a temp solution. The original plan was to enable the USE flag by
default in the chromeos-chrome ebuild, see CL:415324. However, this brought
down the PFQ twice since the change landed without actually being compiled
(it didn't uprev the ebuild) and later on broke the PFQ when an unrelated
change upreved Chrome. See chromium:672119 and chromium:670430 for details.
Adding the dependency in the target-chromium-os ebuild forces the change to
be picked up immediately. Once authpolicy builds and is stable, this CL is
reverted and the original plan is executed.

BUG=chromium:667696
TEST=build_packages on x86-generic and amd64-generic

Change-Id: I6e151082790c15090b6963d5102846a67ee84db9
Reviewed-on: https://chromium-review.googlesource.com/418200
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>