Turn authpolicy on by default for Chrome, mask it out for arm
Some dependencies don't compile on arm (yet). The original plan was to turn
authpolicy on for some boards only. However, we decided to turn it on
for all x86-based boards, so we hit problems rather sooner than later.

This is basically a reland of CL:416049 and CL:415730.

BUG=chromium:667696
TEST=Used emerge to display dependencies, made sure logic for toggling works.

Change-Id: I26dc740ab9262405f1ea4d937c6eac6bb97e62ac
Reviewed-on: https://chromium-review.googlesource.com/419436
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>