Change the depenency of system-locales.
The system-locales requires exactly the same version as glibc. However,
we could not upgrade glibc and system-locales in the same time. This is
because the glibc is built by the sdk bot, but the system-locales is
built for the target. If we upgrade these two package at the same time,
in the commit queue buildbot, the buildbot will try to build glibc for
target and it will fail. So glibc and system-locales can fall out of
sync during an upgrade.

BUG=chromium:674336
TEST=cbuildbot.

Change-Id: Iacfde8063914472b540661f1cd29faabc4905047
Reviewed-on: https://chromium-review.googlesource.com/419696
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>