chromeos-fonts: only generate the cache when building an image
When running build_packages, there's not really a need to have an
up-to-date font cache in the sysroot as it's only used by runtime
tools.  So skip generating the cache then and only do it when we
are running build_image (to generate the actual image).

If someone were to run tools that used fontconfig inside of the
sysroot, they might start seeing slower behavior, but since we
don't have any tools doing that today, it shouldn't be a problem.

BUG=chromium:205424
TEST=`emerge-nyan chromeos-fonts` is now fast
TEST=`./build_image --board nyan` is still slow for this package

Change-Id: I4a89e7f0e85b7c774c555c5818b88c5b4043f578
Reviewed-on: https://chromium-review.googlesource.com/422627
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>