eclass/arc-build: Make library dependencies available from target sysroot
Currently prebuilt Android libraries, including their pkgconfig files
and the pkg-config wrapper itself are located outside of the target
sysroot. This poses a problem when using localy built libraries as
dependencies of other packages is desired.

This patch adds arc-build and arc-build-nyc ebuilds that install
pkgconfig files, pkg-config wrapper and a symlink to the prebuilt
binaries to /opt/google/containers/android, so that all the build
prerequisites are available within the target sysroot.

BUG=b:29343358
TEST=emerge-samus arc-mesa pulls chromeos-base/arc-build

Change-Id: I1d493d0c9526536449722c18a6afccacfafc9d0f
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/422812
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>