cros-workon: Respect CROS_WORKON_SUBDIRS_BLACKLIST for non local fetch.
Before this CL, CROS_WORKON_SUBDIRS_BLACKLIST was only respected when
fetch_method=local. In effect, this meant that the blacklist was
respected for *-9999.ebuild but not for the non -9999 counterpart.

This CL brings the support closer to complete -- we still don't (and
can't) respect the option when we decide to just symlink the already
checked-out source of the developer.

BUG=chromium:676503
TEST=Following passes:
$ cros_workon --board guado_moblab stop chromeos-base/autotest-server
$ emerge-guado_moblab chromeos-base/autotest-server
$ cros_workon --board guado_moblab start chromeos-base/autotest-server
$ emerge-guado_moblab chromeos-base/autotest-server
CQ-DEPEND=CL:424159

complete.

Change-Id: I13f3ec74568c47139d9481678c47df8dfdfe0248
Reviewed-on: https://chromium-review.googlesource.com/422446
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>