cros-workon: Add some incompatible options validation.
There is no way to enforce the sub-directory blacklist if we never copy
sources from the user's checkout. This is OK, since this is not the
intended use case for the CROS_WORKON_SUBDIR_BLACKLIST option.
Let's just validate the ebuilds to be sure.

BUG=chromium:676503
TEST=pre-cq

Change-Id: I6097c449a88977b81bf781a489b46856ce8be851
Reviewed-on: https://chromium-review.googlesource.com/423187
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>