touch_updater: st_flash is conditional dependency
This CL adds the chromeos-base/st_flash package as a conditional
dependency of touch_updater.  This is so that you can set
INPUT_DEVICES="st" for Chromebooks that use st input devices and
therefore need st_flash to be included in their images for firmware
updates.

BUG=chrome-os-partner:61088
TEST=build packages for eve with the correct INPUT_DEVICES line in its
config and st_flash was automatically added in.

Change-Id: I2c7d8e966d9461c757b7600bf600e70f36751e27
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/423212
Commit-Ready: Tai-hsu Lin <sheckylin@google.com>
Tested-by: Tai-hsu Lin <sheckylin@google.com>
Reviewed-by: Tai-hsu Lin <sheckylin@google.com>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>