Revert "cr50-update: show warning screen before reboot"
This reverts commit 7d8db632f5558ba13ee57e3b7be28f7cb5a07696.

There are problems with showing the update splash screen, and with
controlling frecon after that. A better way to do that is being
researched.

The problem can be shows up when switching between normal and dev
modes. To make sure that the change is picked up the ebuild version is
actually bumped up.

BUG=chrome-os-partner:61103

Change-Id: I33b9b60dbc6e4196318b5a64cf986c3f9dcf8220
Reviewed-on: https://chromium-review.googlesource.com/424954
Commit-Ready: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>