[GCC] Fix building gcc with USE='next_gcc'.
A change committed upstream in Feb. 2016 is causing the next_gcc
emerges to fail.  This CL applies a patch to GCC, when building
with USE='next_gcc', that undoes the upstream fix.  This is
intended to be a temporary workaround while we investigate and fix
the actual problem with the upstream change.

BUG=chromium:673427
TEST=Emerged gcc, using USE='next_gcc' with this CL.

Change-Id: I284ab9cfcebc3afb03dcee1b7d2790da41debb36
Reviewed-on: https://chromium-review.googlesource.com/424958
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>