chromeos-ec: Support unified builds
Support building the EC for multiple models in a single board. This uses
the unified build approach. The list of models to build is defined by the
EC_FIRMWARE_UNIBUILD and EC_FIRMWARE_EXTRA_UNIBUILD variables.

BUG=chromium:680151
BRANCH=none
TEST=USE="quiet unibuild" emerge-reef chromeos-ec
Look in build/reef/firmware/ and see that firmware exists for reef, pyro,
snappy. So far it is untested (until I get some more hardware).

Change-Id: Ifce063859442fd32955daabf8e7d919c0af056e7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/427381
Tested-by: Simon Glass <sjg@google.com>
Reviewed-by: Randall Spangler <rspangler@chromium.org>