sys-boot/coreboot: Remove --ignore-existing from coreboot-private rsync
This patch removes the --ignore-existing option from the rsync command
that syncs files from /build/<board>/firmware/coreboot-private into the
coreboot source tree before building. This is important for cases where
we actually want to replace files in coreboot (or subprojects), such as
a binary HDCP firmware blob in RK3399's ARM TF which is uploaded as a
binary to ARM's upstream repository but which we would like to build
from private sources for security reasons.

BUG=chrome-os-partner:59557
TEST=None

Change-Id: I9246e9121486a4c75df0291798c1453752d7374e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/428128
Reviewed-by: Aaron Durbin <adurbin@chromium.org>