libva-intel-driver: Fix GPU HANG on h264 apprtc loopback
One buffer domain was not configured correctly because the buffer is unmapped,
which causes the content to not be synchronized under this test scenario. Then
GPU hang is triggered randomly.  After three hours of testing, GPU hang is not reproduced.

The driver was not freeing the auxiliary batch buffer when we were setting the access domain
for the PAK_OBJ command buffer.  At the same time, the alignment was wrong and should have been
at the cache line boundary.

Finally, fix the order of operations for the pipline flush on SKL+.

BUG=chromium-os-partner:60407
TEST=on Intel emerge and with apprtc as described in bug

Change-Id: I064021be8a46324b47a7eba2a2468e3eaedd9b96
Signed-off-by: Sean V Kelley <seanvk@posteo.de>
Reviewed-on: https://chromium-review.googlesource.com/429030
Commit-Ready: Sean V Kelley <sean.v.kelley@intel.com>
Tested-by: Sean V Kelley <sean.v.kelley@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>