coreboot, libpayload: remove build left overs manually
"make clean" takes while to clean up the tree completely (in case of
coreboot, incl the payload directories) due to more complicated
discovery of what needs cleaning.

Instead, just remove the things that get in the way of _our_ builds, ie.
the content of the build directory.

In case of coreboot, also move objutil in its own directory so kconfig
and various other parts aren't built twice.

This shaves off 1 minute from
$ emerge-kevin libpayload depthcharge chromeos-bmpblk chromeos-ec coreboot chromeos-bootimage
on my system (with a hot filesystem cache), from 4:23 to 3:20.

BUG=chromium:630451
TEST=run with and without change, measure speed improvement

Change-Id: I86276956cb2cf3cc945c2e923ea8813ea9c2e61c
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430710
Reviewed-by: Julius Werner <jwerner@chromium.org>