coreboot: Use variables for the config file names
Rather than repeating the name each time, use variables. This will allow
us to supply a different name in a future CL, which is needed for unified
builds.

I have avoided adding quotes since it is unnecessary and harder to read.
But I can add them in if needed.

BUG=chromium:682781
BRANCH=none
TEST=emerge-reef --nodeps coreboot; see that files are still installed

Change-Id: I86063f8598ef53010f1bc8eb39bac4963b7083ab
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430228
Reviewed-by: Chris Ching <chingcodes@chromium.org>