coreboot: Pass the board config to make_coreboot()
Rather than moving the config into place and then calling make_coreboot(),
pass the required config to the function and have it deal with the details
of how to press the config into use.

Adjust make_coreboot() to store the config in the build directory, so that
it can be installed correctly after all builds are complete.

These changes make it easier to handle unified builds.

BUG=chromium:682781
BRANCH=none
TEST=emerge-reef --nodeps coreboot; see that files are still installed
$ grep  CONFIG_SOC_UART_DEBUG /build/reef/firmware/coreboot.config
$ grep  CONFIG_SOC_UART_DEBUG /build/reef/firmware/coreboot_serial.config
CONFIG_SOC_UART_DEBUG=y

Change-Id: I8427f3bd5d17cb0b693131509c523484983e5419
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430327