Require chromeos-config in chromeos-login
When unified builds is active some packages need the Chrome OS
master configuration. Rather than adding it as an overall Chrome OS
dependency, we have decided to add it to each ebuild that needs it.

Add a dependency for chromeos-login.

BUG=chromium:683360
TEST=USE=unibuild equery-reef g chromeos-login
See that chromeos-config shows up as a dependency

Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Ib28db8828e967f57e05297577ce9725c2e37196d
Reviewed-on: https://chromium-review.googlesource.com/431130
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>