depthcharge: Use a function to avoid repeating options
The same options are used four times. Create a function to avoid repeating
this code. This means we pass these variables to all emake stages, rather
than just those that build the code. This will allow us to pass obj=<path>
(in a future commit) and have it do the right thing.

BUG=chromium:685706
BRANCH=none
TEST=emerge-reef --nodeps depthcharge;
$ find /build/reef/firmware -name "depth*" -or -name "netboot*" -or -name "dev*" -or -name .config | sort
/build/reef/firmware/depthcharge
/build/reef/firmware/depthcharge.config
/build/reef/firmware/depthcharge/depthcharge.elf
/build/reef/firmware/depthcharge/depthcharge.elf.map
/build/reef/firmware/depthcharge/dev.elf
/build/reef/firmware/depthcharge/dev.elf.map
/build/reef/firmware/depthcharge/netboot.elf
/build/reef/firmware/depthcharge/netboot.elf.map

Change-Id: Icfc6f31cf2df033207f36763be215bb0de9bf539
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/433522
Reviewed-by: Aaron Durbin <adurbin@chromium.org>