depthcharge: Move building into a separate function
With unified builds we want to build multiple images, one for each
supported model. In preparation for this, move building of the images into
a separate function.

Drop the 'tc-getCC' call at the same time. It outputs the toolchain but
does not set the CC environment variable, so does not seem useful.

Add a few comments as well.

BUG=chromium:685706
BRANCH=none
TEST=emerge-reef --nodeps depthcharge;
$ find /build/reef/firmware -name "depth*" -or -name "netboot*" -or -name "dev*" -or -name .config | sort
/build/reef/firmware/depthcharge
/build/reef/firmware/depthcharge.config
/build/reef/firmware/depthcharge/depthcharge.elf
/build/reef/firmware/depthcharge/depthcharge.elf.map
/build/reef/firmware/depthcharge/dev.elf
/build/reef/firmware/depthcharge/dev.elf.map
/build/reef/firmware/depthcharge/netboot.elf
/build/reef/firmware/depthcharge/netboot.elf.map

Change-Id: I6945f5345d2f19c3c0dd60eca2087b6e55f2bb7a
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/433524
Reviewed-by: Aaron Durbin <adurbin@chromium.org>