harfbuzz: Drop cairo, glib, and introspection.
Unset the cairo, glib, and introspection USE flags when
building media-libs/harfbuzz. chromeos-chrome and freetype
are the only packages depending on harfbuzz, and neither of
them need it to be built with these flags, which pull in
additional dependencies.

Also unset the cairo USE flag when building
sci-visualization/gnuplot (used by autotest-server-deps),
since it additionally pulls in pango, which wants harfbuzz
to be built with glib.

BUG=chromium:170408
TEST=did some builds, including guado_moblab-paladin

Change-Id: I09ae067bf85744e8cc4ab2296702df0fe0335e0f
Reviewed-on: https://chromium-review.googlesource.com/433395
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>