chromeos-bootimage: Refactor build_image() to use an image type
Rather than passing in the whole filename to build_image(), pass just the
image type, as appended to the filename. So for "image.dev.bin" we pass
"dev" and for "image.bin" we pass "". This will allow unified builds to
provide a model suffix (in a future CL).

Also add some comments to the function.

BUG=chromium:685670
TEST=emerge-reef chromeos-bootimage
$ find /build/reef/firmware/ -name "image*"
/build/reef/firmware/image.net.bin
/build/reef/firmware/image.bin
/build/reef/firmware/image.dev.bin
/build/reef/firmware/image.serial.bin

Change-Id: Ieb590f9f2ccd87587e738ef9fb155f5c3b52ba07
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/434058
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>